requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
Django==4.2.6
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.2.0
flack==1.3.7
Flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pillow==10.2.0
pypiwin32==223
PythonTurtle==0.3.2
pywin32==306
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
